requirements.txt
pytest==7.4.3
pillow==9.5.0
selenium==4.16.1
webdriver-manager==4.0.1
